422: update Gradient sample

+++ a_mini/projects/MiniAgg/04_Scanline/2_ScanlinePacked8.cs
@@ -91,11 +91,11 @@ namespace PixelFarm.Agg
             else
             {
                 m_covers[m_cover_index] = (byte)cover;
-                last_span_index++;
+                last_span_index++; 
 
                 m_spans[last_span_index].cover_index = m_cover_index++;
                 m_spans[last_span_index].x = (short)x;
-                m_spans[last_span_index].len = (short)(-(int)(len));
+                m_spans[last_span_index].len = (short)(-len);
             }
             last_x = x + len - 1;
         }
